% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{data_generation}
\alias{data_generation}
\title{Generate data set}
\usage{
data_generation(
  row_marginals = rowSums(pvLRT::statin46),
  col_marginals = colSums(pvLRT::statin46),
  lambda_mat,
  signal_position,
  omega_vec = rep(0, length(col_marginals))
)
}
\arguments{
\item{row_marginals, col_marginals}{A (possibly named) vector of row and column marginal totals. Must add up to the same total. If named, the names are passed on to the randomly generated matrix/matrices.}

\item{lambda_mat}{A numeric matrix of dimension length(row_marginals) x length(col_marginals). The (i, j)-th entry of lambda_mat determines the signal strength of the i-th adverse event and j-th drug pair.}

\item{signal_position}{A matrix of (0,1) to indicate the position of signals.}

\item{omega_vec}{A vector of zero inflation probabilities. Must be of the same length as col_marginals.}
}
\value{
\itemize{
\item k - Critical value for calculating pFDR
\item optim - Indicate if the k value is optimized or specified
}
}
\description{
Generate data set
}
\examples{
\dontrun{


}
}
