% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_correlation_matrix.R
\name{extract_correlation_matrix}
\alias{extract_correlation_matrix}
\title{Extract correlation matrix from pvBayes object}
\usage{
extract_correlation_matrix(obj, par, by_row = FALSE, method = "pearson")
}
\arguments{
\item{obj}{A pvBayes object with lamabda draws}
}
\value{
\itemize{
\item k - Critical value for calculating pFDR.
\item optim - Indicate if the k value is optimized.
}
}
\description{
Extract correlation matrix from pvBayes object
}
\examples{
\dontrun{
library(pvLRT)
data(statin46)
}
}
