% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_discovery_mat.R
\name{extract_discovery_mat}
\alias{extract_discovery_mat}
\title{Extract discovery matrix from pvBayes_est object}
\usage{
extract_discovery_mat(pvbayes_est_obj, ...)
}
\arguments{
\item{pvBayes_est_obj}{A pvBayes_bayes object with lamabda draws}
}
\value{
\itemize{
\item k - Critical value for calculating pFDR.
\item optim - Indicate if the k value is optimized.
}
}
\description{
Extract discovery matrix from pvBayes_est object
}
\examples{
\dontrun{
library(pvLRT)
data(statin46)
}
}
