% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_normal.R
\name{fit_normal}
\alias{fit_normal}
\title{Fit a N(mu, sigma) model to univariate data y}
\usage{
fit_normal(y, method = "sample", ...)
}
\arguments{
\item{y}{input data}

\item{method}{method of posterior approximation. Defaults to MCMC sampling.}

\item{...}{passed to appropriate stan method (sample, variational, or optimize)}
}
\description{
Fit a N(mu, sigma) model to univariate data y
}
