% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFDR.R
\name{pFDR}
\alias{pFDR}
\title{Compute pFDR}
\usage{
pFDR(lambda_draws, test_stat, optim = TRUE, alpha = 0.05, k = NULL)
}
\arguments{
\item{lambda_draws}{MCMC samples obtianed from \code{pvbayes}}

\item{test_stat}{Bayesian estimates}

\item{optim}{logical. Use specified critical value or optimize pFDR(k)=alpha. Default is TRUE, if FALSE then k must be specified}

\item{alpha}{Confidence level s.t. pFDR(k)=alpha}

\item{k}{Critical value}
}
\value{
\itemize{
\item k - Critical value for calculating pFDR
\item optim - Indicate if the k value is optimized or specified
\item pFDR - Positive false discovery rate value
\item BayesTIE - Bayesian type-I error
}
}
\description{
Compute pFDR
}
\examples{
\dontrun{
library(pvLRT)
data(statin46)
}
}
