% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvBayes-package.R
\docType{package}
\name{pvBayes-package}
\alias{pvBayes}
\alias{pvBayes-package}
\title{pvBayes: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Xinwei Huang \email{xinweihuangstat@gmail.com}

Authors:
\itemize{
  \item Saptarshi Chakraborty
}

}
\keyword{internal}
