% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvbayes.R
\name{pvbayes}
\alias{pvbayes}
\title{Fitting Poisson model with Bayesian methods for pharmacovigilance}
\usage{
pvbayes(
  contin_table,
  model = "zip_horseshoe",
  stan_obj = NULL,
  starting = "random",
  stan_seed = 123,
  stan_chains = 4,
  stan_iter_sampling = 1000,
  stan_cov_rate = 0.5,
  retry = 10,
  stan_parallel_chains = getOption("mc.cores", 4),
  ...
)
}
\arguments{
\item{contin_table}{IxJ contingency table showing pairwise counts of adverse events for I AE (along the rows) and J Drugs (along the columns). Colnames and rownames are needed.}

\item{model}{Options of model:
\itemize{
\item \code{zip_horseshoe} - zero-inflated poisson model with horseshoe prior
\item \code{zip_horseshoe_LKJ} - zero-inflated poisson model under independent assumption with horseshoe prior
}}

\item{stan_obj}{User defined model in stan with input: \code{n} counts in each cells, \code{E} expected value for each cells, and \code{N} total number of cells IxJ}

\item{starting}{option of HMC start point
\itemize{
\item \code{random} - random starting point chosen by cmdstanr
\item \code{LRT} - likelihood ratio estimator by pvLRT
\item \code{MAP}- maximum posterior estimator
}}

\item{stan_seed}{A seed for the (P)RNG to pass to CmdStan.}

\item{stan_chains}{The number of Markov chains to run in CmdStan. The default is 2.}

\item{stan_iter_sampling}{An integer of total sampling number.}

\item{retry}{An integer for maximum repeated number of Stan sampling process.}

\item{...}{additional parameters, currently is not used.}
}
\value{
\itemize{
\item E - Contingency table of expected counts.
\item draws - A list contains the MCMC samples for each parameters.
\item MCMC_convergence - Information of Stan sampler.
}
}
\description{
Fitting Poisson model with Bayesian methods for pharmacovigilance
}
\examples{
\dontrun{
library(pvLRT)
data(statin46)
}
}
