% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvbayes_est.R
\name{pvbayes_est}
\alias{pvbayes_est}
\title{Bayes estimate using MCMC samples}
\usage{
pvbayes_est(lambda_draws, test_method = "mean", prob = NULL, alpha = 0.05)
}
\arguments{
\item{lambda_draws}{MCMC samples obtianed from \code{pvbayes}}

\item{test_method}{The choice of test statistic based on posterior distribution}

\item{alpha}{Confidence level s.t. pFDR(k)=alpha}
}
\value{
\itemize{
\item k - Critical value for calculating pFDR
\item optim - Indicate if the k value is optimized or specified
\item pFDR - Positive false discovery rate value
\item BayesTIE - Bayesian type-I error
}
}
\description{
Bayes estimate using MCMC samples
}
\examples{
\dontrun{
library(pvLRT)
data(statin46)
}
}
