% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvbayes_long.R
\name{pvbayes_long}
\alias{pvbayes_long}
\title{Fitting Poisson model with Bayesian methods for pharmacovigilance}
\usage{
pvbayes_long(
  contin_table,
  model = "horseshoe",
  stan_obj = NULL,
  stan_seed = 123,
  stan_chains = 2,
  stan_parallel_chains = 2,
  ...
)
}
\description{
Fitting Poisson model with Bayesian methods for pharmacovigilance
}
