% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvBayes_setup.R
\name{pvBayes_setup}
\alias{pvBayes_setup}
\title{Set up stan compiled objects for pvBayes package}
\usage{
pvBayes_setup(...)
}
\arguments{
\item{...}{additional arguments passed to cmdstanr::cmdstan_model()}
}
\description{
Set up stan compiled objects for pvBayes package
}
