% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_generation.R
\name{signal_generation}
\alias{signal_generation}
\title{Generate signal matrix}
\usage{
signal_generation(
  I,
  J,
  sig_position,
  sig_strength,
  non_sig_corr_mat = matrix(0, nrow = J - 1, ncol = J - 1) \%>\% `diag<-`(1),
  sig_corr_mat = NULL,
  var_vec = rep(1e-04, J - 1)
)
}
\arguments{
\item{I}{An integer of row number.}

\item{J}{An integer of column number.}

\item{sig_position}{A list of pairs (i, j) where signals are placed. To specify the entirety i-th row (or j-th column) use c(i, 0) (or c(0, j)).}

\item{sig_strength}{A list of signal strength corresponding to the position specified in \code{sig_position}.}

\item{non_sig_corr_mat}{A (J-1)x(J-1) matrix. Correlation matrix of multivariate-normal distribution for non-signal cells generation.}

\item{sig_corr_mat}{A (J-1)x(J-1) matrix. Correlation matrix of multivariate-normal distribution for signal cells generation.}

\item{var_vec}{A positive numeric vector length by (J-1). Variance vector of multivariate-normal for both signal and non-signal generation. Suggested to be relatively small value.}

\item{sig_mean_vec}{A positive numeric vector length by (J-1). Mean vector of multivariate-normal distribution for non-signal cells generation.}
}
\value{
\itemize{
\item lambda - A matrix of generated signals.
\item signal_true - A matrix of (0,1) to indicate the position of signals.
}
}
\description{
Generate signal matrix
}
\examples{
\dontrun{
library(pvLRT)
data(statin46)
}
}
