% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_output_to_rvar.R
\name{stan_output_to_rvar}
\alias{stan_output_to_rvar}
\title{Basis function for transform MCMC samples to rvar object}
\usage{
stan_output_to_rvar(obj, ...)
}
\arguments{
\item{obj}{a stan object produced by "pvbayes"}

\item{...}{additional parameters}
}
\description{
Basis function for transform MCMC samples to rvar object
}
