% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_draws_list.R
\name{extract_draws_list}
\alias{extract_draws_list}
\title{Extract lambda draws from pvBayes object and transform to a list form}
\usage{
extract_draws_list(obj, par = "lambda")
}
\arguments{
\item{obj}{A pvBayes object with lambda draws}

\item{par}{Name of draws transform to list}
}
\value{
\itemize{
\item A list of lambda draws, each element is one MCMC sample.
}
}
\description{
Extract lambda draws from pvBayes object and transform to a list form
}
\examples{
\dontrun{
library(pvLRT)
data(statin46)
}
}
