% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvbayes.R
\name{pvbayes}
\alias{pvbayes}
\title{Fitting Poisson model with Bayesian methods for pharmacovigilance}
\usage{
pvbayes(
  contin_table,
  model = "horseshoe",
  stan_obj = NULL,
  stan_seed = 123,
  stan_chains = 2,
  stan_parallel_chains = 2,
  ...
)
}
\arguments{
\item{contin_table}{IxJ contingency table showing pairwise counts of adverse events for I AE (along the rows) and J Drugs (along the columns)}

\item{model}{Select Bayesian methods
\itemize{
\item \code{horseshoe} - model 1
\item \code{beta_prime} - model 2
}}

\item{stan_obj}{User defined model in stan with input: \code{n} counts in each cells, \code{E} expected value for each cells, and \code{N} total number of cells IxJ}

\item{stan_seed}{A seed for the (P)RNG to pass to CmdStan.}

\item{stan_chains}{The number of Markov chains to run in CmdStan. The default is 2.}

\item{stan_parallel_chains}{The maximum number of MCMC chains to run in parallel in CmdStan.}

\item{...}{additional parameters}
}
\value{
\itemize{
\item lambda_draws - A tibble contain the MCMC samples for each poisson rate parameter.
\item contin_table_long - The converted contingency table in long format tibble.
}
}
\description{
Fitting Poisson model with Bayesian methods for pharmacovigilance
}
\examples{
\dontrun{
library(pvLRT)
data(statin46)
mod <- pvbayes(contin_table = statin46, model = "horseshoe")

#obtain the MCMC samples
mod$lambda_draws
}
}
