% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvbayes_est.R
\name{pvbayes_est}
\alias{pvbayes_est}
\title{Signal detection function using MCMC samples with pDFR threshold}
\usage{
pvbayes_est(
  pvbayes_obj,
  test_stat = function(x) {
     quantile(x, 0.05)
 },
  alpha = 0.05,
  thresh = 1.1,
  ...
)
}
\arguments{
\item{test_stat}{A function of test statistic (e.g. mean, quantile).}

\item{alpha}{A real value between (0,1). Confidence level s.t. pFDR(k)=alpha.}

\item{lambda_draws}{An rvar object of MCMC samples obtained from \code{pvbayes}.}
}
\value{
\itemize{
\item k - Critical value for calculating pFDR.
\item optim - Indicate if the k value is optimized.
\item pFDR - Positive false discovery rate.
\item test_stat - A matrix of test statistics in each cell.
\item BayesTIE - A matrix of Bayesian type-I error.
\item range_test_stat - A vector of maximal and minimal test statistics.
\item sig_pfdr - A matrix of discoveries (0 or 1) using pFDR as the threshold.
\item sig_naive - A reference matrix of discoveries (0 or 1) using .05 quantile.
}
}
\description{
Signal detection function using MCMC samples with pDFR threshold
}
\examples{
\dontrun{
library(pvLRT)
data(statin46)
}
}
