% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvbayes_est.R
\name{pvbayes_est}
\alias{pvbayes_est}
\title{Bayes estimate using MCMC samples}
\usage{
pvbayes_est(lambda_draws, est_quantile, alpha = 0.05)
}
\arguments{
\item{lambda_draws}{MCMC samples obtianed from \code{pvbayes}}

\item{est_quantile}{Bayesian estimates using pth-quantile of the MCMC samples}

\item{alpha}{Confidence level s.t. pFDR(k)=alpha}
}
\value{
\itemize{
\item k - Critical value for calculating pFDR
\item optim - Indicate if the k value is optimized or specified
\item pFDR - Positive false discovery rate value
\item BayesTIE - Bayesian type-I error
}
}
\description{
Bayes estimate using MCMC samples
}
